spark/requirements.txt
pyspark==3.4.3
delta-spark==2.4.0
importlib_metadata>=4.0.0
